Seperated GUI_Main into GUI_Main and GUI_MainScreen.py. There are only GUI classes is that file. Executing will be only on GUI_Main.py GUI_Main파일을 GUI_Main과 GUI_MainScreen으로 나눴습니다. GUI_MainScreen은 화면을 구성하는 클래스만 있고, 실행은 GUI_Main에서 실행될 것입니다.
